Export instances tree lookup from project-build
This utility is needed to filter loaded data per page on published
sites. We aleady have it in builder so easy to move.
