+++ apps/builder/app/shared/tree-utils.test.ts
@@ -16,7 +16,6 @@ import {
   findClosestRichTextInstance,
   findParentInstance,
   findSubtreeLocalStyleSources,
-  findTreeInstances,
   getInstanceAncestorsAndSelf,
   insertInstanceMutable,
   insertInstancesCopyMutable,
@@ -437,24 +436,6 @@ test("find parent instance", () => {
       { type: "id", value: "5" },
     ],
   });
-});
-
-test("find all tree instances", () => {
-  const instances: Instances = new Map([
-    createInstancePair("1", "Body", [{ type: "id", value: "3" }]),
-    // this is outside of subtree
-    createInstancePair("2", "Box", []),
-    // these should be matched
-    createInstancePair("3", "Box", [
-      { type: "id", value: "4" },
-      { type: "id", value: "5" },
-    ]),
-    createInstancePair("4", "Box", []),
-    createInstancePair("5", "Box", []),
-    // this one is from other tree
-    createInstancePair("6", "Box", []),
-  ]);
-  expect(findTreeInstances(instances, "3")).toEqual(new Set(["3", "4", "5"]));
 });
 
 test("insert tree of instances copy and provide map from ids map", () => {
